+++ R/arg_rep.R
@@ -10,7 +10,7 @@
 #' \dontrun{
 #' arg_rep(lm)
 #' 
-#' clipr::write_clip(c('formula, data, subset, weights, na.action, method = "qr"',    
+#' clipr::write_clip(c('formula, data, subset, weights, na.action, method = "qr",',    
 #'     'model = TRUE, x = FALSE, y = FALSE, qr = TRUE, singular.ok = TRUE,',
 #'     'contrasts = NULL, offset, ...'
 #' ))